# Conflicts:
#	src/main/java/ca/mcgill/ecse223/quoridor/QuoridorApplication.java

#	src/main/java/ca/mcgill/ecse223/quoridor/controller/Quoridor223Controller.java

#	src/test/java/ca/mcgill/ecse223/quoridor/features/CucumberStepDefinitions.java

+++ src/test/java/ca/mcgill/ecse223/quoridor/features/CucumberStepDefinitions.java
@@ -1,10 +1,11 @@
 package ca.mcgill.ecse223.quoridor.features;
 
 import java.sql.Time;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import ca.mcgill.ecse223.quoridor.QuoridorApplication.*;
+import ca.mcgill.ecse223.quoridor.QuoridorApplication;
 import ca.mcgill.ecse223.quoridor.controller.Quoridor223Controller;
 import ca.mcgill.ecse223.quoridor.model.Board;
 import ca.mcgill.ecse223.quoridor.model.Direction;
@@ -22,19 +23,13 @@ import ca.mcgill.ecse223.quoridor.model.WallMove;
 import cucumber.api.PendingException;
 import io.cucumber.java.After;
 import io.cucumber.java.en.And;
+import io.cucumber.java.en.But;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import io.cucumber.java.en.But;
+import junit.framework.Assert;
 
 public class CucumberStepDefinitions {
-
-	private Quoridor quoridor;
-	private Board board;
-	private Player player1;
-	private Player player2;
-	private Player currentPlayer;
-	private Game game;
 
 	// ***********************************************
 	// Background step definitions
@@ -48,21 +43,24 @@ public class CucumberStepDefinitions {
 
 	@Given("^The game is running$")
 	public void theGameIsRunning() {
-		theGameIsNotRunning();
-		createAndStartGame();
+		initQuoridorAndBoard();
+		ArrayList<Player> createUsersAndPlayers = createUsersAndPlayers("user1", "user2");
+		createAndStartGame(createUsersAndPlayers);
 	}
 
 	@And("^It is my turn to move$")
 	public void itIsMyTurnToMove() throws Throwable {
-		currentPlayer = player1;
-		Quoridor223Application.getQuoridor().getCurrentGame().getCurrentPosition().setPlayerToMove(currentPlayer);
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
+		Player currentPlayer = quoridor.getCurrentGame().getWhitePlayer();
+		QuoridorApplication.getQuoridor().getCurrentGame().getCurrentPosition().setPlayerToMove(currentPlayer);
 	}
 
 	@Given("The following walls exist:")
 	public void theFollowingWallsExist(io.cucumber.datatable.DataTable dataTable) {
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
 		List<Map<String, String>> valueMaps = dataTable.asMaps();
 		// keys: wrow, wcol, wdir
-		Player[] players = { player1, player2 };
+		Player[] players = { quoridor.getCurrentGame().getWhitePlayer(), quoridor.getCurrentGame().getBlackPlayer() };
 		int playerIdx = 0;
 		int wallIdxForPlayer = 0;
 		for (Map<String, String> map : valueMaps) {
@@ -70,8 +68,8 @@ public class CucumberStepDefinitions {
 			Integer wcol = Integer.decode(map.get("wcol"));
 			// Wall to place
 			// Walls are placed on an alternating basis wrt. the owners
-			Wall wall = Wall.getWithId(playerIdx * 10 + wallIdxForPlayer);
-
+			//Wall wall = Wall.getWithId(playerIdx * 10 + wallIdxForPlayer);
+			Wall wall = players[playerIdx].getWall(wallIdxForPlayer); // above implementation sets wall to null
 			String dir = map.get("wdir");
 
 			Direction direction;
@@ -85,13 +83,13 @@ public class CucumberStepDefinitions {
 			default:
 				throw new IllegalArgumentException("Unsupported wall direction was provided");
 			}
-			new WallMove(0, 1, players[playerIdx], board.getTile((wrow - 1) * 9 + wcol - 1), game, direction, wall);
+			new WallMove(0, 1, players[playerIdx], quoridor.getBoard().getTile((wrow - 1) * 9 + wcol - 1), quoridor.getCurrentGame(), direction, wall);
 			if (playerIdx == 0) {
-				game.getCurrentPosition().removeWhiteWallsInStock(wall);
-				game.getCurrentPosition().addWhiteWallsOnBoard(wall);
+				quoridor.getCurrentGame().getCurrentPosition().removeWhiteWallsInStock(wall);
+				quoridor.getCurrentGame().getCurrentPosition().addWhiteWallsOnBoard(wall);
 			} else {
-				game.getCurrentPosition().removeBlackWallsInStock(wall);
-				game.getCurrentPosition().addBlackWallsOnBoard(wall);
+				quoridor.getCurrentGame().getCurrentPosition().removeBlackWallsInStock(wall);
+				quoridor.getCurrentGame().getCurrentPosition().addBlackWallsOnBoard(wall);
 			}
 			wallIdxForPlayer = wallIdxForPlayer + playerIdx;
 			playerIdx++;
@@ -101,9 +99,14 @@ public class CucumberStepDefinitions {
 
 	}
 
-	@And("I do not have a wall in my hand")
+	@And("I shall not have a wall in my hand")
 	public void iDoNotHaveAWallInMyHand() {
-		// Walls are in stock for all players
+		// GUI-related feature -- TODO for later
+	}
+	
+	@And("^I have a wall in my hand over the board$")
+	public void iHaveAWallInMyHandOverTheBoard() throws Throwable {
+		// GUI-related feature -- TODO for later
 	}
 
 	// ***********************************************
@@ -117,91 +120,149 @@ public class CucumberStepDefinitions {
 	 * are implemented
 	 * 
 	 */
-	@And("I have a wall in my hand over the board")
-	public void iHaveAWallInMyHandOverTheBoard() {
-		throw new PendingException();
-	}
 
-	@Given("A wall move candidate exists with (.*) at position ((.*), (.*))")
-	public void wallMoveCandidateExists(String dir, int row, int col) {
-		getWallMoveCandidate(dir, row, col);
+	/**
+	 * @author Le-Li Mao
+	 */
+	@Given("A wall move candidate exists with {string} at position \\({int}, {int})")
+	public void aWallMoveCandidateExistsWithAtPosition(String orientation, int row, int col) {
+//		setWall(orientation,row, col);
+		
 	}
-
-	@And("The wall candidate is at the (.*) edge of the board")
-	public void Thewallcandidateisattheedgeoftheboard(String side) {
-		throw new PendingException();
+	/**
+	 * @author Le-Li Mao
+	 */
+	@And("A wall move candidate shall exist with {string} at position \\({int}, {int})")
+	public void aWallMoveCandidateShallExistWithDirAtPositionNrowNcol(String orientation, int row, int col) {
+		Assert.assertEquals(true, hasWallCandidate(orientation,row,col));
 	}
-
-	@And("The wall candidate is not at the (.*) edge of the board")
-	public void Thewallcandidateisnotattheedgeoftheboard(String side) {
-		throw new PendingException();
+	/**
+	 * @author Le-Li Mao
+	 */
+	@Given("The wall move candidate with {string} at position \\({int}, {int}) is valid")
+	public void theWallMoveCandidateWithDirAtPositionRowColIsValid(String orientation, int row, int col) {
+//		setWall(orientation,row, col);
 	}
-
-	@When("I try to move the wall (.*)")
-	public void iTryToMoveTheWall(String side) {
-		Quoridor223Controller.moveWall(side);
-	}
-
-	@Then("The wall shall be moved over the board to position ((.*), (.*))")
-	public void theWallShallBeMovedOverTheBoardToPosition(int row, int col) {
-		throw new PendingException();
-	}
-
-	@And("A wall move candidate shall exist with (.*) at position ((.*), (.*))")
-	public void aWallMoveCandidateShallExist(String dir, int row, int col) {
-		throw new PendingException();
-	}
-
-	@Then("I should be notified that my move is illegal")
-	public void IShouldBeNotifiedThatMyMoveIsIllegal() {
-		// Feel like this is GUI
-		throw new PendingException();
-	}
-
-	@Given("The wall move candidate with (.*) at position ((.*), (.*)) is valid")
-	public void theWallMoveCandidateIsValid(String dir, int row, int col) {
-		setWall(dir, row, col);
-	}
-
-	@Given("The wall move candidate with (.*) at position ((.*), (.*)) is invalid")
-	public void TheWallMoveCandidateWithIsInvalid(String dir, int row, int col) {
-		setWall(dir, row, col);
-	}
-
+	/**
+	 * @author Le-Li Mao
+	 */
 	@When("I release the wall in my hand")
 	public void iReleaseTheWallInMyHand() {
-		Quoridor223Controller.dropWall();
-	}
+		try{
+			Quoridor223Controller.dropWall();
+		}
+		catch (Exception e){
 
-	@But("A wall move is registered with (.*) at position ((.*), (.*))")
-	public void aWallMoveIsRegistered() {
+		}
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@But("A wall move shall be registered with {string} at position \\({int}, {int})")
+	public void aWallMoveIsRegisteredWithDirAtPositionRowCol(String orientation, int col, int row) {
 		throw new PendingException();
 	}
-
-	@And("My move is completed")
+	/**
+	 * @author Le-Li Mao
+	 */
+	@And("My move shall be completed")
 	public void myMoveIsCompleted() {
 		throw new PendingException();
 	}
-
-	@And("It is not my turn to move")
+	/**
+	 * @author Le-Li Mao
+	 */
+	@And("It shall not be my turn to move")
 	public void itIsNotMyTurnToMove() {
 		throw new PendingException();
 	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@Given("The wall move candidate with {string} at position \\({int}, {int}) is invalid")
+	public void theWallMoveCandidateWithDirAtPositionRowColIsInvalid(String orientation, int col, int row) {
+		throw new PendingException();
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@But("No wall move shall be registered with {string} at position \\({int}, {int})")
+	public void noWallMoveIsRegisteredWithDirAtPositionRowCol(String orientation, int col, int row) {
+		throw new PendingException();
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@And("The wall candidate is not at the {string} edge of the board")
+	public void theWallCandidateIsNotAtTheSideEdgeOfTheBoard() {
+		throw new PendingException();
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@Then("The wall shall be moved over the board to position \\({int}, {int})")
+	public void theWallShallBeMovedOverTheBoardToPositionNrowNcol(int col, int row) {
+		Quoridor quoridor =QuoridorApplication.getQuoridor();
+		Game curGame = quoridor.getCurrentGame();
+		WallMove candidate = curGame.getWallMoveCandidate();
+		Tile tile= candidate.getTargetTile();
+		boolean reached = isCordEqual(row,col,tile.getRow(),tile.getColumn());
+		Assert.assertEquals(true,reached);
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@When("I try to move the wall {string}")
+	public void iTryToMoveTheWallSide(String side) {
+		try {
+			Quoridor223Controller.moveWall(side);
+		}
+		catch (Exception e){
 
-	@Then("I shall be notified that my wall move is invalid")
+		}
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@And("The wall candidate is at the {string} edge of the board")
+	public void theWallCandidateIsAtTheSideEdgeOfTheBoard(String side) {
+		WallMove candidate = QuoridorApplication.getQuoridor().getCurrentGame().getWallMoveCandidate();
+		int row = candidate.getTargetTile().getRow();
+		int col = candidate.getTargetTile().getColumn();
+		if(side.equalsIgnoreCase("UP")){
+			Assert.assertEquals(1,row);
+		}
+		else if (side.equalsIgnoreCase("DOWN")){
+			Assert.assertEquals(8,row);
+		}
+		else if (side.equalsIgnoreCase("RIHGT")){
+			Assert.assertEquals(8,col);
+		}
+		else {
+			Assert.assertEquals(1,col);
+		}
+	}
+	/**
+	 * @author Le-Li Mao
+	 */
+	@Then("I shall be notified that my move is illegal")
+	public void iShouldBeNotifiedThatMyMoveIsIllegal() {
+		//GUI
+		throw new PendingException();
+	}
+	@Then ("I shall be notified that my wall move is invalid")
 	public void iShallBeNotifiedThatMyWallMoveIsInvalid() {
-		// Feel like this is GUI
+		throw new PendingException();
 	}
-
-	@But("No wall move is registered with (.*) at position ((.*), (.*))")
-	public void noWallMoveIsRegistered(String direction, int row, int col) {
-
+	@And("I shall have a wall in my hand over the board")
+	public void iShallHaveAWallInMyHandOverTheBoard() {
+		//GUI 
+		throw new PendingException();
 	}
-	
-	
-	
-	
-	
+	@And("It shall be my turn to move")
+	public void itShallBeMyTurnToMove() {
+		
+	}
 	// ***********************************************
 	// Clean up
 	// ***********************************************
@@ -209,8 +270,18 @@ public class CucumberStepDefinitions {
 	// After each scenario, the test model is discarded
 	@After
 	public void tearDown() {
-		quoridor.delete();
-		quoridor = null;
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
+		// Avoid null pointer for step definitions that are not yet implemented.
+		if (quoridor != null) {
+			quoridor.delete();
+			quoridor = null;
+		}
+		for (int i = 0; i < 20; i++) {
+			Wall wall = Wall.getWithId(i);
+			if(wall != null) {
+				wall.delete();
+			}
+		}
 	}
 
 	// ***********************************************
@@ -220,8 +291,8 @@ public class CucumberStepDefinitions {
 	// Place your extracted methods below
 
 	private void initQuoridorAndBoard() {
-		quoridor = Quoridor223Application.getQuoridor();
-		board = new Board(quoridor);
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
+		Board board = new Board(quoridor);
 		// Creating tiles by rows, i.e., the column index changes with every tile
 		// creation
 		for (int i = 1; i <= 9; i++) { // rows
@@ -231,7 +302,8 @@ public class CucumberStepDefinitions {
 		}
 	}
 
-	private void createUsersAndPlayers(String userName1, String userName2) {
+	private ArrayList<Player> createUsersAndPlayers(String userName1, String userName2) {
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
 		User user1 = quoridor.addUser(userName1);
 		User user2 = quoridor.addUser(userName2);
 
@@ -244,9 +316,10 @@ public class CucumberStepDefinitions {
 		 * __________ | | | | |x-> <-x| | | |__________|
 		 * 
 		 */
-		// @formatter:on
-		player1 = new Player(new Time(thinkingTime), user1, 9, Direction.Horizontal);
-		player2 = new Player(new Time(thinkingTime), user2, 1, Direction.Horizontal);
+
+		//@formatter:on
+		Player player1 = new Player(new Time(thinkingTime), user1, 9, Direction.Horizontal);
+		Player player2 = new Player(new Time(thinkingTime), user2, 1, Direction.Horizontal);
 
 		Player[] players = { player1, player2 };
 
@@ -257,20 +330,28 @@ public class CucumberStepDefinitions {
 				new Wall(i * 10 + j, players[i]);
 			}
 		}
+		
+		ArrayList<Player> playersList = new ArrayList<Player>();
+		playersList.add(player1);
+		playersList.add(player2);
+		
+		return playersList;
 	}
 
-	private void createAndStartGame() {
+	private void createAndStartGame(ArrayList<Player> players) {
+		Quoridor quoridor = QuoridorApplication.getQuoridor();
 		// There are total 36 tiles in the first four rows and
 		// indexing starts from 0 -> tiles with indices 36 and 36+8=44 are the starting
 		// positions
-		Tile player1StartPos = board.getTile(36);
-		Tile player2StartPos = board.getTile(44);
+		Tile player1StartPos = quoridor.getBoard().getTile(36);
+		Tile player2StartPos = quoridor.getBoard().getTile(44);
+		
+		Game game = new Game(GameStatus.Running, MoveMode.PlayerMove, players.get(0), players.get(1), quoridor);
 
-		PlayerPosition player1Position = new PlayerPosition(player1, player1StartPos);
-		PlayerPosition player2Position = new PlayerPosition(player2, player2StartPos);
+		PlayerPosition player1Position = new PlayerPosition(quoridor.getCurrentGame().getWhitePlayer(), player1StartPos);
+		PlayerPosition player2Position = new PlayerPosition(quoridor.getCurrentGame().getBlackPlayer(), player2StartPos);
 
-		game = new Game(GameStatus.Running, MoveMode.PlayerMove, player1, player2, quoridor);
-		GamePosition gamePosition = new GamePosition(0, player1Position, player2Position, player1, game);
+		GamePosition gamePosition = new GamePosition(0, player1Position, player2Position, players.get(0), game);
 
 		// Add the walls as in stock for the players
 		for (int j = 0; j < 10; j++) {
@@ -284,27 +365,32 @@ public class CucumberStepDefinitions {
 
 		game.setCurrentPosition(gamePosition);
 	}
-
-	/**
-	 * Create select the wall and set it to the wall candidate in the game
-	 * 
-	 * @param dir
-	 * @param row
-	 * @param col
-	 */
-	private void getWallMoveCandidate(String dir, int row, int col) {
-		// create a new WallMove Candidate and place the corresponding tile
-		Direction wallDirection = dir.equals("horizontal") ? Direction.Horizontal : Direction.Vertical;
-		game.setWallMoveCandidate(new WallMove(0, 1, player1, board.getTile((row - 1) * 9 + (col - 1)), game,
-				wallDirection, game.getCurrentPosition().getBlackWallsInStock(0)));
+	private void getWallMoveCandidate(Player player, String dir, int row, int col) {
+		//create a new WallMove Candidate and place the corresponding tile
+		Board board = QuoridorApplication.getQuoridor().getBoard();
+		Game game =  QuoridorApplication.getQuoridor().getCurrentGame();
+		Direction wallDirection = dir.equals("horizontal")?Direction.Horizontal:Direction.Vertical;
+		game.setWallMoveCandidate(new WallMove(0, 1, player, board.getTile((row-1)*9+(col-1)), game, wallDirection, game.getCurrentPosition().getBlackWallsInStock(0)));
 	}
-
-	private void setWall(String dir, int row, int col) {
-		// 25 is a valid position of 3 row and 8th column
-		if (game.getWallMoveCandidate() != null) {
-			game.getWallMoveCandidate().setTargetTile(board.getTile((row - 1) * 9 + (col - 1)));
-		} else {
-			getWallMoveCandidate(dir, row, col);
+	private void setWall(Player player, String dir, int row, int col) {
+		Game game =  QuoridorApplication.getQuoridor().getCurrentGame();
+		Board board = QuoridorApplication.getQuoridor().getBoard();
+		if(game.getWallMoveCandidate()!=null) {
+		game.getWallMoveCandidate().setTargetTile(board.getTile((row-1)*9+(col-1)));
 		}
+		else {
+			getWallMoveCandidate(player,dir, row, col);
+		}
+	}
+	public boolean hasWallCandidate(String dir, int row, int col){
+		Game game =  QuoridorApplication.getQuoridor().getCurrentGame();
+		if(game.getWallMoveCandidate()==null)return false;
+		WallMove candidate = game.getWallMoveCandidate();
+		if(candidate.getTargetTile().getRow()!=row || candidate.getTargetTile().getColumn()!=col)return false;
+		if(candidate.getWallDirection()!=(dir.equalsIgnoreCase("horizontal")?Direction.Horizontal:Direction.Vertical))return false;
+		return true;
+	}
+	public boolean isCordEqual(int row1,int col1, int row2, int col2){
+		return row1==row2 && col1==col2;
 	}
 }
